Put Godot integers as Long in Bundles

+++ src/main/java/org/godotengine/androidplugin/firebase/Utils.java
@@ -152,7 +152,7 @@ public class Utils {
         if (value instanceof Boolean) {
             bundle.putBoolean(key, (Boolean) value);
         } else if (value instanceof Integer) {
-            bundle.putInt(key, (Integer) value);
+            bundle.putLong(key, Long.valueOf((Integer) value));
         } else if (value instanceof Double) {
             bundle.putDouble(key, (Double) value);
         } else if (value instanceof String) {
